create todo state hook
